requirements.txt
flask==3.0.2
requests==2.31.0
gitpython==3.1.42
humanize==4.9.0
